+++ app.html
@@ -42,7 +42,7 @@
 		<svg id="editor" xmlns="http://www.w3.org/2000/svg"></svg><svg id="movePreview"></svg><svg
 			id="overPreview"></svg><svg id="snapping"></svg>
 	</div>
-	<div class="tools">
+	<div class="tools hide">
 		<div class="propertyScrubber">
 			<div class="scrub"></div>
 		</div>
